% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_eval.R
\name{compute_set_reachability}
\alias{compute_set_reachability}
\title{Get reachability for set containing xopt}
\usage{
compute_set_reachability(mole_trace, xopt)
}
\arguments{
\item{mole_trace}{}

\item{xopt}{}
}
\description{
Get reachability for set containing xopt
}
