% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHVArchive.R
\name{computeHVArchive}
\alias{computeHVArchive}
\title{Hypervolume for all prefixes of an archive of points}
\usage{
computeHVArchive(archive, ref_point)
}
\arguments{
\item{archive}{Numeric matrix with one pair of objective values per column,
and two rows}

\item{ref_point}{Numeric vector of length two containing the reference point.}
}
\value{
A vector of length \code{ncol(archive)} containing the hypervolume of all
prefixes of the archive.
}
\description{
Implements an efficient (\eqn{n \log n}) algorithm to compute the hypervolume
for all prefixes of an archive of objective vectors.
}
\details{
Currently only supports \strong{bi-objective problems}.
Assumes \strong{minimization}.

The "historical" hypervolume of an archive is particularly interesting for
benchmarking purposes.
}
