% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbob_biobj.R
\name{generateBiObjBBOBData}
\alias{generateBiObjBBOBData}
\title{Generate data for bi-objective BBOB functions}
\usage{
generateBiObjBBOBData(dimensions, fid, iid)
}
\arguments{
\item{dimensions}{number of dimensions for the test problem}

\item{fid}{function ID (1-55) of the bi-objective BBOB problem}

\item{iid}{instance ID (1-15) of the bi-objective BBOB problem}
}
\value{
A list containing the following elements:
\itemize{
\item \code{fn}: \code{smoof} function of the requested bi-objective BBOB function
\item \code{opt_f1}: Optimal input value for the first constituent BBOB function
\item \code{opt_f2}: Optimal input value for the second constituent BBOB function
\item \code{ref_point}: Reference point for this function as used by the bi-objective BBOB for evaluation
\item \code{ideal_point}: Ideal point for this function as used by the bi-objective BBOB for evaluation
}
}
\description{
Generate data for bi-objective BBOB functions
}
\examples{
# generate data for the 1st instance of the 10th function in 2d
bbob_biobj_data <- generateBiObjBBOBData(dimensions = 2L, fid = 10L, iid = 1L)
}
