% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeAsparFunction}
\alias{makeAsparFunction}
\title{\code{smoof} generator for the Aspar problem with 2-3 dimensions and objectives.}
\usage{
makeAsparFunction(dimensions = 2L, n.objectives = 2L)
}
\arguments{
\item{dimensions}{Number of dimensions (2-3)}

\item{n.objectives}{Number of dimensions (2-3)}
}
\value{
\code{smoof} function of the configured Aspar problem
}
\description{
\code{smoof} generator for the Aspar problem with 2-3 dimensions and objectives.
}
\examples{
fn <- makeAsparFunction(dimensions = 2L, n.objectives = 2L)
}
