% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeBiObjMPM2Function}
\alias{makeBiObjMPM2Function}
\title{\code{smoof} generator for bi-objective MPM2 problems}
\usage{
makeBiObjMPM2Function(
  dimensions = 2,
  n.peaks.1 = 3,
  topology.1 = "random",
  seed.1 = 4,
  n.peaks.2 = 3,
  topology.2 = "random",
  seed.2 = 8
)
}
\arguments{
\item{dimensions}{Number of search space dimensions}

\item{n.peaks.1}{Number of peaks for first MPM2 problem}

\item{topology.1}{Type of topology for first MPM2 problem}

\item{seed.1}{Random seed for first MPM2 problem}

\item{n.peaks.2}{Number of peaks for second MPM2 problem}

\item{topology.2}{Type of topology for second MPM2 problem}

\item{seed.2}{Random seed for second MPM2 problem}
}
\value{
\code{smoof} function of the configured bi-objective MPM2 function
}
\description{
\code{smoof} generator for bi-objective MPM2 problems
}
\examples{
fn <- makeBiObjMPM2Function(dimensions = 2,
  n.peaks.1 = 3, topology.1 = "random", seed.1 = 4,
  n.peaks.2 = 3, topology.2 = "random", seed.2 = 8
)
}
