% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSGKFunction}
\alias{makeSGKFunction}
\title{\code{smoof} generator for the SGK problem}
\usage{
makeSGKFunction()
}
\value{
\code{smoof} function containing the bi-objective SGK problem
}
\description{
\code{smoof} generator for the SGK problem
}
\examples{
fn <- makeSGKFunction()
}
