% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSimpleBiSphereFunction}
\alias{makeSimpleBiSphereFunction}
\title{`smoof` generator for simple Bi-Sphere problem}
\usage{
makeSimpleBiSphereFunction()
}
\value{
\code{smoof} function containing a simple Bi-Sphere problem with
centers \code{c(-1, -1)} and \code{c(1, 1)}
}
\description{
`smoof` generator for simple Bi-Sphere problem
}
\examples{
fn <- makeSimpleBiSphereFunction()
}
