% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runif_box}
\alias{runif_box}
\title{Generates a vector distributed uniformly at random between \code{lower} and
\code{upper}}
\usage{
runif_box(lower, upper)
}
\arguments{
\item{lower}{lower bounds of search space}

\item{upper}{upper bounds of search space}
}
\value{
Vector of length \code{length(lower)} with entries uniformly at
random between \code{lower} and \code{upper} per dimension
}
\description{
Generates a vector distributed uniformly at random between \code{lower} and
\code{upper}
}
\examples{
# Create a vector of length 2 with elements uniformly at random between -5 and 5
runif_box(c(-5, -5), c(5, 5))

# Generating a matrix of values
points <- lapply(1:nstarts, function(x) runif_box(lower, upper))
points_matrix <- do.call(rbind, points)
}
