% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeAsparFunction}
\alias{makeAsparFunction}
\title{`smoof` generator for the Aspar problem with 2-3 dimensions and objectives.}
\usage{
makeAsparFunction(dimensions = 2L, n.objectives = 2L)
}
\value{
\code{smoof} function of the configured Aspar problem
}
\description{
`smoof` generator for the Aspar problem with 2-3 dimensions and objectives.
}
\examples{
fn <- makeAsparFunction(dimensions = 2L, n.objectives = 2L)
}
