% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeBiRosenbrockFunction}
\alias{makeBiRosenbrockFunction}
\title{`smoof` generator for Bi-Rosenbrock function}
\usage{
makeBiRosenbrockFunction()
}
\value{
\code{smoof} function containing the Bi-Rosenbrock function
}
\description{
`smoof` generator for Bi-Rosenbrock function
}
\examples{
fn <- makeBiRosenbrockFunction()
}
